backend/requirements.txt
fastapi 
uvicorn 
python-multipart 
PyMuPDF 
google-generativeai
load_dotenv